+++ api/plan-dft-1d.c
@@ -21,8 +21,7 @@
 #include "api.h"
 #include "dft.h"
 
-X(plan) X(plan_dft_1d)(unsigned long n, C *in, C *out, int sign, 
-		       unsigned int flags)
+X(plan) X(plan_dft_1d) (int n, C *in, C *out, int sign, int flags)
 {
-     return X(plan_dft)(1, &n, in, out, sign, flags);
+     return X(plan_dft) (1, &n, in, out, sign, flags);
 }
